requirements.txt
streamlit
pillow
requests
xlwt
httpx
asyncio
feedparser
telethon

pandas
pymorphy2
numpy
matplotlib
nltk 
